requirements.txt
websockets
python-dotenv
requests
Flask